Merged PR 11846297: Private Stamps + Settings Refactor
Add bicep template which devs can use to deploy a private environment
Includes an Azure App Config instance so that you can launch the project without having to go to portal and grab a bunch of settings

Since I was already touching a bunch of settings, I refactored how we touch/ inject settings so that we share the same code across all the projects

Apologies if this breaks anyone in the near-term; these files change daily and I needed to go ahead and get things in. I tried to verify that at least all projects start up.

----
#### AI description  (iteration 1)
#### PR Classification
Code refactor and feature addition.

#### PR Summary
This pull request refactors settings and introduces private stamps for Kusto clusters.
- Refactored `aca-kusto.json` to nest Kusto cluster configurations under an "External" key.
- Deleted `AppSettings.cs` and `KustoServiceClient.cs` files.
- Added `KustoClientService.cs` to handle Kusto client operations with various authentication methods.
- Updated `AppSettings.cs` to streamline and reorganize settings, including the addition of `CosmosDBSettings` and `LoggingSettings`.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->