Merged PR 11870408: Clean up the 1P & 1P.ACA project.
Combine the IcmAutomationClient into ICMWorkflowClient
The IcmPlugin and ICMPlugin still need to combine, the work is half done for now.
Make the AppConfig only be used in the development mode.
Unify the model class between 1P and 1P.ACA project.
Fix the setting of external AppSetting is not loading (FirstPartyAgent.Core\Extensions\FirstPartyAgentsConfiguration.cs)