Merged PR 12222098: Setting React chat client to load under the root path of Agent site
Previously we hid the new client under /react/, but now you should be able to hit the new experience by hitting the root path.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request modifies the routing configuration to load the React chat client under the root path of the Agent site and includes a minor code improvement.
- Changes in `src/Agent/Agent.Web/Program.cs` to map the fallback to the React client.
- Changes in `src/Agent/Agent.Web/Client/src/src/Space/Activities/ThreadsMenu.tsx` to handle potential null values in thread messages.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->