Merged PR 12255502: Fix dashboard url for build scenario of traffic analysis
Fix dashboard url.
![image.png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/e8ebd6d6-4f3f-4a98-ac01-ac8297162f89/pullRequests/12255502/attachments/image.png)
Unable to reproduce the auth error, but the dashboard url is wrong, making the analysis flow broken.

It can now take snapshot of the correct dashboard and make useful analysis.
![image (2).png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/e8ebd6d6-4f3f-4a98-ac01-ac8297162f89/pullRequests/12255502/attachments/image%20%282%29.png)
![image (3).png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/e8ebd6d6-4f3f-4a98-ac01-ac8297162f89/pullRequests/12255502/attachments/image%20%283%29.png)

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to correct the dashboard URL in the build scenario of traffic analysis.

#### PR Summary
This pull request addresses a bug in the `GraphDBPlugin.cs` file where the incorrect property was used for the dashboard URL configuration.
- `GraphDBPlugin.cs`: Corrected the property from `resource.ResourceType` to `resource.SubscriptionId` for the `var-sub` variable in the URL configuration.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->