Merged PR 12298668: Switch from Working Memory Set to Private Bytes for Azure Monitor Memory Metrics
Working Memory set is subject to spikes
The Private bytes metric is more accurate when it comes to the memory consumption by the app

----
#### AI description  (iteration 1)
#### PR Classification
Code modification to switch memory metrics in Azure Monitor.

#### PR Summary
This pull request updates the memory metrics used in Azure Monitor from "MemoryWorkingSet" to "PrivateBytes" for more accurate monitoring. It also includes a minor typo correction.
- In `AppInsightsPlugin.cs`, corrected a typo in the variable name from `instrumentionKey` to `instrumentationKey`.
- In `MetricsPlugin.cs`, replaced the metric "MemoryWorkingSet" with "PrivateBytes" and removed a commented-out line related to "MemoryWorkingSet".
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->