Allow user to ask LocalAuthAgent to suppress warnings (#146)
This allows the user to tell the system to ignore a resource for a
specified time, so as not to get pinged by the scanner over and over.

It does so by creating a new graph node attached the resource in
question that records various options for that resource.

Here's what that looks like to the user (the warning about format
mismatch has since been fixed):

![image](https://github.com/user-attachments/assets/2ab4a01f-d7f8-4897-b8e5-7d87b903b4eb)

---------

Co-authored-by: Avi Pilosof <avip@microsoft.com>
Co-authored-by: Casey Hulbert <chulbert@microsoft.com>