Implement Emerging Issue handling (#246)
This PR is adding feature of Emerging Issue Handling. The objective is
to mitigate the Emerging issue by seeing the match of the emerging issue
incident.

* `EmergingIssuePlugin` that will detect the emerging issue
* `EmergingIssueManagerPlugin` that will manage the emerging issue. It
will pre-processing the emerging issue and extract the summarized
information.

Talked with Nazeef, for the first version we go with the chat window for
both EmergingIssueManager and EmergingIssue Handler.


## EmegingIssueManager
Emerging Issue Manager accept commands that start with `/` for example,
`/register <incidentId>` That enables `/register` and `/deregister` and
`/list_all` commands that filters emerging issue.


![image](https://github.com/user-attachments/assets/8f1a579b-8889-4506-b4f7-3fd371bd1306)

## EmergingIssue Handler

Detecting Emerging Issue for CRI. by providing the IndicentId. Currently
LSI is not supported.
Current configuration optimizing Functions. Since it introduce
generalization, We can introduce AppService wide handling is available.
(I'm asking the case of AppService wide Emerging issue)


![image](https://github.com/user-attachments/assets/72cb7fdf-4a0e-4bb6-aa2e-883b2260b8fa)

Next Step: 
* Merge PR
* Support AppServiceWide Filtering

---------

Co-authored-by: Tsuyoshi Ushio <ushio@simplearchitect.com>