fix: variable `defaultHandler` is not proper set in `CreateMessage` API for incident handler thread (#1098)
### Description
https://github.com/serverless-paas-balam/sreagent-runtime/pull/1019
added a variable called `defaultHandler` to manage whether to use
custom-instruction or the agent framework.

Currently, this variable is not properly set in the `CreateMessage` API,
which may cause the incident handler thread to unexpectedly switch to
the agent framework when a user adds a message to the chat.

### Type of Change
- [x] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test coverage improvement
- [ ] Prompt change (requires approval from prompt author/owner)


### Risk
- [ ] High – Errors could cause MAJOR regression of many scenarios.
(Example: new large features or high-level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios.
(Example: somewhat complex bug fixes, small new features)
- [x] Small – No issues are expected. (Example: Very small bugfixes,
string changes, or configuration settings changes)